Ajout supprimer+ confirmation du choix

+++ TD4MGDB/app/routes/developers.js
@@ -1,14 +1,16 @@
 import Route from '@ember/routing/route';
 
 export default Route.extend({
-    
+
 model(){
     return this.store.findAll('developer');
 },
 
 actions:{
     delete(dev){
+      if( confirm("La suppression est définitive, êtes-vous sûr ?") ){
         dev.destroyRecord();
+      }
     }
 }
 });
